Runtime Manager and Vector 3 Fatih

+++ Squirrel-Engine/include/Vector2.h
@@ -5,7 +5,7 @@ namespace SRMath
 	{
 	public:
 		Vector2();
-		Vector2(int x, int y);
+		Vector2(float x, float y);
 		float x;
 		float y;
 	};
