+++ Squirrel-Engine/src/Math/Vector2.cpp
@@ -8,7 +8,7 @@ namespace SRMath
 		y = 0;
 	}
 
-	Vector2::Vector2(int _x, int _y)
+	Vector2::Vector2(float _x, float _y)
 	{
 		x = _x;
 		y = _y;
